+++ src/cfd_lib/headers/utils.hpp
@@ -7,15 +7,15 @@
 #include "SFML/System.hpp"
 #include "SFML/Graphics.hpp"
 
-uint8_t map_to_range(uint8_t i_start, uint8_t i_end, uint8_t o_start, uint8_t o_end, uint8_t value);
+int map_to_range(int i_start, int i_end, int o_start, int o_end, int value);
 float map_to_range(float i_start, float i_end, float o_start, float o_end, float value);
 sf::Color hsv(float hue, float saturation, float value, float alpha);
 
 class GridInfo{
     public:
-        uint16_t row_size;
-        uint16_t col_size;
-        uint32_t grid_size;
+        int row_size;
+        int col_size;
+        int grid_size;
 
         int indx(int x, int y) const;
 };
